refactor(infer): update imports to use new core/adapter architecture
